requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
phidata
google-generativeai
duckduckgo-search
pytube
yt_dlp
duckduckgo-search
pydantic-ai==0.0.20
pydantic-graph
openai
colorama==0.4.6
dataclasses==0.6
asyncio==3.4.3
devtools==0.12.2
html-to-markdown==1.1.0
beautifulsoup4
tavily-python
google-generativeai